consolidate RulePickerLine-related components
