+++ js/app/delivery/pricing/components/RulePickerLine/PickerIsLoading.js
